sage.rings.function_field.derivations: Split out from .maps
